Adds a Tool for using the Transfer MkDir Action Provider
This implements a slightly different, but intending to be compatible,
version of some elements of a tool. Notably:

- Properties for the state can be set on the `__init__` method of the
tool class. This leads to a model where tools tend to be included in
clients based on instances rather than Class objects or class name
strings. However, because all arguments to `__init__` have default
values, and those defaults are using the pattern of other tools, it
should behave as other Tools when used based on Class object or name.

- The `flow_definition` is dynamically generated based on parameters
to the `__init__`. This allows some of the things often set by
`modifiers` to be set by the values in `__init__`.

- Likewise, the `required_input` is dynamically generated and can
adapt to different values placed in the `__init__`.