fix: Add recursive option to publish v2
This problem seems odd to me, but transfer _should_ automatically detect
recursive transfers and not require the extra boolean. It's odd to me why
this suddenly started failing. Adding the extra option by default True does
fix this for the common use-case where the entity being transferred is a directory.